Unit tests for configurations
